[WIP] Proof-of-concept separate redux store for index
