+++ src/Index/reduxer.jsx
@@ -10,13 +10,13 @@ import { components } from '../params';
 
 export const ReduxIndexBarChart = (() => {
   const mapStateToProps = (state) => {
-    if (state.homepage && state.homepage.projectsByName) {
+    if (state.index && state.index.projectsByName) {
       const projectList = Object.values(
-        state.homepage.projectsByName,
+        state.index.projectsByName,
       ).sort(sortCompare);
       return {
         projectList,
-        countNames: state.homepage.countNames,
+        countNames: state.index.countNames,
       };
     }
     return {};
@@ -30,13 +30,13 @@ export const ReduxIndexBarChart = (() => {
 
 export const ReduxIndexCounts = (() => {
   const mapStateToProps = (state) => {
-    if (state.homepage && state.homepage.projectsByName) {
+    if (state.index && state.index.projectsByName) {
       const projectList = Object.values(
-        state.homepage.projectsByName,
+        state.index.projectsByName,
       ).sort(sortCompare);
       return {
         projectList,
-        countNames: state.homepage.countNames,
+        countNames: state.index.countNames,
       };
     }
     return {};
